requirements.txt
fastapi
uvicorn
gunicorn
pytest
faker
pydantic
pydantic-settings
redis
aiomysql
aiosqlite
SQLAlchemy
alembic
starlette
httpx
socksio
pytest-asyncio
ruff
python-multipart
pyjwt
passlib
